requirements.txt
biopython==1.78
imageio==2.9.0
jsonlines==2.0.0
numpy==1.21.1
pandas==1.3.1
PyYAML==5.4.1
scipy==1.7.1
torch==1.7.0+cu110
tensorboardX==2.4
tqdm==4.62.0
yarg==0.1.9
